# D3d12: Fixes