--- a/rpcs3/Emu/RSX/D3D12/D3D12Formats.h
+++ b/rpcs3/Emu/RSX/D3D12/D3D12Formats.h
@@ -97,3 +97,8 @@ DXGI_FORMAT get_index_type(u8 index_type) noexcept;
  * Convert vertex attribute format and size to DXGI_FORMAT
  */
 DXGI_FORMAT get_vertex_attribute_format(u8 type, u8 size) noexcept;
+
+/**
+ * Convert scissor register value to D3D12_RECT
+ */
+D3D12_RECT get_scissor(u32 horizontal, u32 vertical) noexcept;