--- a/rpcs3/Emu/RSX/D3D12/D3D12GSRender.cpp
+++ b/rpcs3/Emu/RSX/D3D12/D3D12GSRender.cpp
@@ -351,14 +351,7 @@ void D3D12GSRender::end()
 	};
 	getCurrentResourceStorage().command_list->RSSetViewports(1, &viewport);
 
-	D3D12_RECT box =
-	{
-		0,
-		0,
-		(LONG)clip_w,
-		(LONG)clip_h,
-	};
-	getCurrentResourceStorage().command_list->RSSetScissorRects(1, &box);
+	getCurrentResourceStorage().command_list->RSSetScissorRects(1, &get_scissor(rsx::method_registers[NV4097_SET_SCISSOR_HORIZONTAL], rsx::method_registers[NV4097_SET_SCISSOR_VERTICAL]));
 
 	getCurrentResourceStorage().command_list->IASetPrimitiveTopology(get_primitive_topology(draw_mode));