This PR fixes some defects found by investigating SH3. Game shouldn't lack geometry.
The alpha test sample background is also correctly displayed and some Braid glitches (cloud overlapping geometry) are fixed.