@danilaml 2015-11-10
Latest merges to master have degraded my performance in dice elf. Also I think after d3d12 as a lib pr rpcs3 crashes on game close, but It doesn't even correctly exit and still stays as hidden process holding resources and can be killed only through cmd taskkill /f