@vlj 2015-11-10
Will investigate the performance regression in dice