@danilaml 2015-11-10
@vlj can you look at crashing as well (I get it 100% of the time in closing of dice elf)? It makes testing harder.