@danilaml 2015-11-10
@vlj git bisect told me that https://github.com/RPCS3/rpcs3/commit/e76abf875f60585f622c0ff6ded239625d6b8f5a
is the first bad commit