@vlj 2015-11-11
@Zangetsu38 you need to increase buffer texture size, looks for
	"m_textureUploadData.init(m_device.Get(), 1024 * 1024 * 512, D3D12_HEAP_TYPE_UPLOAD, D3D12_HEAP_FLAG_NONE);" and use a bigger value than 512.