@Zangetsu38 2015-11-11
@vlj Thx, works now with 1024