@vlj 2015-11-11
The crash when exiting dice seems to be older than this commit, it looks like a crash happens in wxWidget, when deleting keyboard or mouse classes.