@danilaml 2015-11-11
@vlj ah, sorry, I din't make it clear - this commit https://github.com/RPCS3/rpcs3/commit/e76abf875f60585f622c0ff6ded239625d6b8f5a is for performance issues.
As for crashes - it started appearing for me around the time of d312 separation/projects/dynlib conundrum. Also I usually see some dx12 destructors in the callstack. Like ~RenderTargets or something.