@danilaml 2015-11-11
@vlj nevermind the crashes issue, it's most likely cause by configurations PR by @O1L (since it affects oGL in the same way too).
Still it leaves the performance issues (it totally destroys my iGPU performance (60 -> 3 fps) and my dGPU takes a noticable hit (37 -> 27 fps))