@vlj 2015-11-12
Rebased